// src/utils/types.ts
export interface RPGSystem {
    gameTitle: string;
    versionId: number;
    locale: string;
    partyMembers: number[];
    currencyUnit: string;
    startMapId: number;
    startX: number;
    startY: number;
    variables: string[];
    switches: string[];
    terms?: {
        basic: string[];
        commands: string[];
        params: string[];
        messages: Record<string, string>;
    };
    [key: string]: unknown;
}